refactor(components|shared): move bundles into shared to resolve circular dependency | hj
