+++ packages/components/stencil.config.ts
@@ -1,5 +1,5 @@
 import * as path from 'node:path';
-import type { TagName } from '@porsche-design-system/shared';
+import { bundles } from '@porsche-design-system/shared';
 import replace from '@rollup/plugin-replace';
 import type { Config } from '@stencil/core';
 import { version } from './package.json';
@@ -19,30 +19,6 @@ process.env.PATH = `${fakeNpmPath}:${process.env.PATH}`;
 
 const isDevBuild = process.env.PDS_IS_STAGING === '1';
 
-// specify chunking of components that can't be used standalone
-// it's important to list the parent component first since it affects the chunk name
-export const bundles: { components: TagName[] }[] = [
-  { components: ['p-drilldown', 'p-drilldown-item', 'p-drilldown-link'] },
-  { components: ['p-multi-select', 'p-multi-select-option'] },
-  { components: ['p-segmented-control', 'p-segmented-control-item'] },
-  { components: ['p-radio-group', 'p-radio-group-option'] },
-  { components: ['p-select', 'p-select-option'] },
-  { components: ['p-stepper-horizontal', 'p-stepper-horizontal-item'] },
-  {
-    components: [
-      'p-table',
-      'p-table-body',
-      'p-table-head',
-      'p-table-head-row',
-      'p-table-head-cell',
-      'p-table-row',
-      'p-table-cell',
-    ],
-  },
-  { components: ['p-tabs', 'p-tabs-item'] },
-  { components: ['p-text-list', 'p-text-list-item'] },
-  { components: ['p-toast', 'p-toast-item'] },
-];
 
 export const config: Config = {
   namespace: 'porsche-design-system',
