+++ packages/shared/package.json
@@ -21,9 +21,7 @@
     "test:unit": "vitest --config tests/unit/config/vitest.config.ts",
     "clean:build": "rm -rf ./src/lib && rm -rf ./dist"
   },
-  "dependencies": {
-    "@porsche-design-system/component-meta": "0.0.0"
-  },
+  "dependencies": {},
   "devDependencies": {
     "@octokit/core": "^7.0.6",
     "change-case": "^5.4.4",
